Extract board device lifecycle emission
