+++ crates/fstart-codegen/src/stage_gen/board_gen.rs
@@ -58,12 +58,14 @@ use fstart_types::{
 use super::tokens::hex_addr;
 
 mod boot_media;
+mod lifecycle;
 mod logger;
 mod model;
 mod phases;
 mod state;
 
 use boot_media::{anchor_bytes_stmt, match_boot_media};
+use lifecycle::{init_all_devices_body, init_device_body};
 use logger::install_logger_body;
 use model::{device_provides, BoardCtx};
 use phases::{phase_init_body, PhaseSpec};
@@ -2557,361 +2559,6 @@ fn payload_load_uefi_body(platform: Platform, ctx: &BoardCtx<'_>) -> TokenStream
 
 /// Emit the body of `Board::init_device`.
 ///
-/// Produces `match id { ... }` with one arm per enabled,
-/// non-structural, non-ACPI device.  Each arm walks the device's
-/// ancestor chain (root-first) and, for every non-inited ancestor
-/// plus the target itself:
-///
-/// 1. Constructs the device via `Device::new(&cfg)` or
-///    `BusDevice::new_on_bus(&cfg, &parent)` (picking based on the
-///    driver's `is_bus_device` metadata), storing the result in
-///    `self.<field>`.
-/// 2. Calls `self.<field>.as_mut().ok_or(_)?.init()?`.
-/// 3. Marks `self._inited.set(id)` so subsequent `init_device` calls
-///    return early via the `if self._inited.contains(id)` guard at
-///    the top of the arm.
-///
-/// The ancestor-chain inlining matches the old `ensure_device_ready`
-/// behaviour: when a capability references a device X, its
-/// construction triggers `.init()` on its whole ancestor chain
-/// root-first (SuperIO needs its LPC southbridge programmed, etc.).
-///
-/// Wildcard arm halts the board — the executor only passes ids
-/// declared in the plan.
-///
-/// # Error propagation
-///
-/// Each `.init()` returns `Result<(), DeviceError>`; the `?` operator
-/// propagates up.  Construction itself (`Device::new`) can also
-/// return `Err`; same pattern.  The trait's `-> Result<(), DeviceError>`
-/// return type matches the old `.unwrap_or_else(halt)` but is
-/// composable — the executor decides the failure policy (currently
-/// always `halt()`, see the `if board.init_device(id).is_err()`
-/// arms in `run_stage`).
-fn init_device_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    use super::config_ser::{config_tokens, driver_type_tokens};
-
-    let entries: Vec<(TokenStream, TokenStream)> =
-        enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-            .filter(|idx| {
-                let inst = &ctx.instances[*idx];
-                !inst.is_structural() && !inst.is_acpi_only()
-            })
-            .map(|idx| {
-                let id_lit = proc_macro2::Literal::u8_unsuffixed(idx as u8);
-                let helper = format_ident!("__fstart_init_device_{}", idx);
-                // Walk root→target collecting every non-structural,
-                // non-acpi-only, enabled device — this matches
-                // `ensure_device_ready`'s chain construction.
-                let chain = chain_from_root(idx, ctx.device_tree, ctx.devices, ctx.instances);
-                let steps = chain.iter().map(|&step_idx| {
-                    let step_dev = &ctx.devices[step_idx];
-                    let step_inst = &ctx.instances[step_idx];
-                    let step_field = format_ident!("{}", step_dev.name.as_str());
-                    let step_id_lit = proc_macro2::Literal::u8_unsuffixed(step_idx as u8);
-                    let ty = driver_type_tokens(step_inst);
-                    let cfg = config_tokens(step_inst);
-                    let cfg_static = format_ident!("__FSTART_CFG_{}", step_idx);
-                    // Construction — dispatch on `is_bus_device`.
-                    //
-                    // Keep construction in per-device out-of-line helpers.  A
-                    // single monolithic `match id` function gets one stack frame
-                    // large enough for the biggest board config on any arm, which
-                    // is too expensive for x86 CAR bootblocks.
-                    let construct = if step_inst.meta().is_bus_device {
-                        let parent_name = walk_to_real_parent(
-                            step_idx,
-                            ctx.device_tree,
-                            ctx.devices,
-                            ctx.instances,
-                        );
-                        match parent_name {
-                            Some(pname) => {
-                                let parent = format_ident!("{}", pname);
-                                quote! {
-                                    static mut #cfg_static: core::mem::MaybeUninit<<#ty as fstart_services::BusDevice>::Config> = core::mem::MaybeUninit::uninit();
-                                    // SAFETY: generated stage init is single-threaded and writes this
-                                    // config before constructing the device. The returned reference is
-                                    // retained for the firmware lifetime.
-                                    let _cfg_ref: &'static <#ty as fstart_services::BusDevice>::Config = unsafe {
-                                        let _cfg_ptr = core::ptr::addr_of_mut!(#cfg_static).cast::<<#ty as fstart_services::BusDevice>::Config>();
-                                        _cfg_ptr.write(#cfg);
-                                        &*_cfg_ptr
-                                    };
-                                    let _parent_ref = this.#parent
-                                        .as_ref()
-                                        .ok_or(fstart_services::device::DeviceError::InitFailed)?;
-                                    let mut _dev = <#ty>::new_on_bus(_cfg_ref, _parent_ref)?;
-                                    _dev.init()?;
-                                    this.#step_field = Some(_dev);
-                                }
-                            }
-                            None => quote! {
-                                static mut #cfg_static: core::mem::MaybeUninit<<#ty as fstart_services::Device>::Config> = core::mem::MaybeUninit::uninit();
-                                // SAFETY: generated stage init is single-threaded and writes this
-                                // config before constructing the device. The returned reference is
-                                // retained for the firmware lifetime.
-                                let _cfg_ref: &'static <#ty as fstart_services::Device>::Config = unsafe {
-                                    let _cfg_ptr = core::ptr::addr_of_mut!(#cfg_static).cast::<<#ty as fstart_services::Device>::Config>();
-                                    _cfg_ptr.write(#cfg);
-                                    &*_cfg_ptr
-                                };
-                                let mut _dev = <#ty>::new(_cfg_ref)?;
-                                _dev.init()?;
-                                this.#step_field = Some(_dev);
-                            },
-                        }
-                    } else {
-                        quote! {
-                            static mut #cfg_static: core::mem::MaybeUninit<<#ty as fstart_services::Device>::Config> = core::mem::MaybeUninit::uninit();
-                            // SAFETY: generated stage init is single-threaded and writes this
-                            // config before constructing the device. The returned reference is
-                            // retained for the firmware lifetime.
-                            let _cfg_ref: &'static <#ty as fstart_services::Device>::Config = unsafe {
-                                let _cfg_ptr = core::ptr::addr_of_mut!(#cfg_static).cast::<<#ty as fstart_services::Device>::Config>();
-                                _cfg_ptr.write(#cfg);
-                                &*_cfg_ptr
-                            };
-                            let mut _dev = <#ty>::new(_cfg_ref)?;
-                            _dev.init()?;
-                            this.#step_field = Some(_dev);
-                        }
-                    };
-                    quote! {
-                        if !this._inited.contains(#step_id_lit) {
-                            #construct
-                            this._inited.set(#step_id_lit);
-                        }
-                    }
-                });
-                let helper_def = quote! {
-                    #[inline(never)]
-                    fn #helper(
-                        this: &mut _BoardDevices,
-                    ) -> Result<(), fstart_services::device::DeviceError> {
-                        if this._inited.contains(#id_lit) {
-                            return Ok(());
-                        }
-                        #(#steps)*
-                        Ok(())
-                    }
-                };
-                let arm = quote! { #id_lit => #helper(self), };
-                (helper_def, arm)
-            })
-            .collect();
-
-    let helper_defs = entries.iter().map(|(helper, _)| helper);
-    let arms = entries.iter().map(|(_, arm)| arm);
-
-    quote! {
-        #(#helper_defs)*
-        match id {
-            #(#arms)*
-            _ => {
-                fstart_log::error!("init_device: unknown device id {}", id);
-                Err(fstart_services::device::DeviceError::InitFailed)
-            }
-        }
-    }
-}
-
-/// Walk a device's ancestor chain root-first, excluding structural
-/// / ACPI-only / disabled ancestors (they have no `self.<field>`).
-///
-/// Returns indices in root-first order ending at `target_idx`.
-/// Mirrors `ensure_device_ready`'s chain-building step in the old
-/// codegen.
-fn chain_from_root(
-    target_idx: usize,
-    device_tree: &[DeviceNode],
-    devices: &[DeviceConfig],
-    instances: &[DriverInstance],
-) -> Vec<usize> {
-    let mut chain = Vec::new();
-    let mut cursor = Some(target_idx);
-    while let Some(idx) = cursor {
-        let inst = &instances[idx];
-        let dev = &devices[idx];
-        if dev.enabled && !inst.is_structural() && !inst.is_acpi_only() {
-            chain.push(idx);
-        }
-        cursor = device_tree[idx].parent.map(|p| p as usize);
-    }
-    chain.reverse();
-    chain
-}
-
-/// Walk an ancestor chain until the first non-structural parent.
-///
-/// Used for `BusDevice::new_on_bus` to pick the *real* parent
-/// reference to pass (skipping structural wrappers like the LPC node
-/// between a SuperIO and its southbridge).  Returns `None` if the
-/// device is a root (no non-structural ancestor).
-fn walk_to_real_parent<'a>(
-    child_idx: usize,
-    device_tree: &[DeviceNode],
-    devices: &'a [DeviceConfig],
-    instances: &[DriverInstance],
-) -> Option<&'a str> {
-    let mut current = device_tree[child_idx].parent?;
-    loop {
-        let idx = usize::from(current);
-        if !instances[idx].is_structural() {
-            return Some(devices[idx].name.as_str());
-        }
-        current = device_tree[idx].parent?;
-    }
-}
-
-/// Emit the body of `Board::init_all_devices`.
-///
-/// Called by the executor's `CapOp::DriverInit` arm to bulk-init
-/// everything not already brought up by an earlier per-capability
-/// init.  Parameters:
-///
-/// - `skip`: devices whose `DriverInit` should be silently skipped
-///   (already initialised in this or a previous stage).
-/// - `gated`: devices whose `.init()` must check the runtime
-///   boot-source register (sunxi `fstart_soc_sunxi::boot_media()`)
-///   against the device's `boot_media_ids`; only init if the
-///   current byte matches.
-///
-/// The body iterates enabled, non-structural, non-ACPI-only devices
-/// in root-first order (matching the old `sorted_indices` order
-/// emitted by `plan_gen`).  For each device:
-///
-/// - `skip.contains(id)` → skip entirely.
-/// - `gated.contains(id)` → emit `if matches!(_bm, VAL_1|VAL_2|..) {
-///   self.init_device(id)? }`.
-/// - otherwise → `self.init_device(id)?`.
-///
-/// Framebuffer devices: the old codegen tracked init success in a
-/// `_<name>_ok: bool` local so UEFI's GOP config could be
-/// conditional.  In the new model `self._inited.contains(fb_id)`
-/// serves the same role.  A failing framebuffer init leaves
-/// `self._inited` un-set and the UEFI builder emits
-/// `framebuffer: None`.
-fn init_all_devices_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    use super::capabilities::boot_media_values_for_device;
-
-    let is_egon = ctx.config.soc_image_format == fstart_types::SocImageFormat::AllwinnerEgon;
-
-    // Collect all runtime-active devices in root-first order.
-    // `enabled_indices` already filters `!dev.enabled / is_acpi_only /
-    // is_structural / excluded`; we just add an ordering step to
-    // match the old `topological_sort` output.  The device tree is
-    // already root-first by construction (`ron_loader::build_device_tree`
-    // assigns depths and emits parent-first), so iterating
-    // `enabled_indices(...)` in insertion order matches.
-    let mut dev_statements = TokenStream::new();
-    let mut has_any_gated = false;
-
-    for idx in enabled_indices(ctx.devices, ctx.instances, ctx.excluded) {
-        let inst = &ctx.instances[idx];
-        if inst.is_structural() || inst.is_acpi_only() {
-            continue;
-        }
-        let dev = &ctx.devices[idx];
-        if ctx.instances[idx].provides(Service::PciRootBus) {
-            continue;
-        }
-        let id_lit = proc_macro2::Literal::u8_unsuffixed(idx as u8);
-        let is_framebuffer = ctx.instances[idx].provides(Service::Framebuffer);
-        // Framebuffer failures are non-fatal (UEFI path still uses
-        // `self._inited.contains(fb_id)` as an availability flag);
-        // other devices halt on failure to match the old codegen.
-        let on_err = if is_framebuffer {
-            quote! {
-                fstart_log::warn!("driver init failed (framebuffer, continuing)");
-                // Do NOT set `_inited` — UEFI config reads
-                // `self._inited.contains(#id_lit)` as the availability
-                // flag.
-            }
-        } else {
-            quote! {
-                fstart_log::error!("FATAL: driver init failed for id {}", #id_lit);
-                fstart_platform::halt();
-            }
-        };
-
-        let bm_values = if is_egon {
-            // Safe to call boot_media_values_for_device — only sunxi
-            // drivers have entries, and it panics on unknown drivers
-            // which we guard against with `catch_unwind`.
-            let dev_name = dev.name.as_str();
-            std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                boot_media_values_for_device(dev_name, ctx.devices, ctx.instances)
-            }))
-            .unwrap_or_default()
-        } else {
-            Vec::new()
-        };
-
-        // Inner init call — runs iff the device is in neither `skip`
-        // nor `gated`, or iff gated and boot_media matches.
-        let init_call = quote! {
-            match self.init_device(#id_lit) {
-                Ok(()) => {}
-                Err(_) => {
-                    #on_err
-                }
-            }
-        };
-
-        let gated_check = if !bm_values.is_empty() && is_egon {
-            has_any_gated = true;
-            let val_lits = bm_values
-                .iter()
-                .map(|v| proc_macro2::Literal::u8_unsuffixed(*v))
-                .collect::<Vec<_>>();
-            quote! {
-                if gated.contains(#id_lit) {
-                    if matches!(_bm, #(#val_lits)|*) {
-                        #init_call
-                    } else {
-                        fstart_log::info!(
-                            "skipping driver init (boot-media gated, not active): id {}",
-                            #id_lit,
-                        );
-                    }
-                } else {
-                    #init_call
-                }
-            }
-        } else {
-            quote! {
-                // Device is not in the gating table; run unconditionally.
-                #init_call
-            }
-        };
-
-        dev_statements.extend(quote! {
-            if !skip.contains(#id_lit) {
-                #gated_check
-            }
-        });
-    }
-
-    // Read sunxi's boot-media byte once up front if any gated arms exist.
-    let bm_preamble = if has_any_gated && is_egon {
-        quote! {
-            // SAFETY: _egon_sram_base is the BROM entry point.
-            let _bm = unsafe {
-                fstart_soc_sunxi::boot_media_at(self._egon_sram_base as usize)
-            };
-        }
-    } else {
-        // Silence unused-var warnings on non-gated stages.
-        quote! { let _ = gated; }
-    };
-
-    quote! {
-        #bm_preamble
-        #dev_statements
-    }
-}
-
 // =======================================================================
 // Helpers
 // =======================================================================
